Move Strategic Base Planning to promo, and update.
There's no strong reason for two.
