+++ src/server/cards/promo/StrategicBasePlanning.ts
@@ -10,9 +10,9 @@ export class StrategicBasePlanning extends PreludeCard {
   constructor() {
     super({
       name: CardName.STRATEGIC_BASE_PLANNING,
-      tags: [Tag.BUILDING],
+      tags: [Tag.CITY, Tag.BUILDING, Tag.SPACE],
 
-      startingMegacredits: -8,
+      startingMegacredits: -3,
 
       behavior: {
         colonies: {buildColony: {}},
@@ -20,16 +20,16 @@ export class StrategicBasePlanning extends PreludeCard {
       },
 
       metadata: {
-        cardNumber: 'P08',
+        cardNumber: 'X65',
         renderData: CardRenderer.builder((b) => {
-          b.minus().megacredits(8).city().colonies();
+          b.minus().megacredits(3).city().colonies();
         }),
-        description: 'Pay 8M€. Place a city. Place a colony.',
+        description: 'Pay 3M€. Place a city. Place a colony.',
       },
     });
   }
   public override bespokePlay(player: IPlayer) {
-    player.stock.deduct(Resource.MEGACREDITS, 8);
+    player.stock.deduct(Resource.MEGACREDITS, 3);
     PathfindersExpansion.addToSolBank(player);
     return undefined;
   }
